refactor(core): move doseToMl para doseUnit e quebra o ciclo de imports
adherenceLogic importava densityFor de doseUnit, e doseUnit passou a importar
doseToMl de adherenceLogic (F4b) — ciclo real. doseToMl só depende de densityFor,
então desce para doseUnit e a seta fica em um sentido só.

Sob Hermes um ciclo assim resolve para undefined em tempo de import conforme a
ordem de avaliação, e o sintoma seria em runtime no mobile, não no tsc.
